Use int32 for positioning system internally
Note that we won't be able to use full range of Int32 data type, since many computations still make use of floating point precision! Int32 is used mostly as an API change, while the core engine still performs many calculations within the floating point range!

+++ src/core/tile.go
@@ -82,7 +82,7 @@ func (o *Object) NewTile() {
 			return
 		}
 
-		dest := rl.NewRectangle(o.Position.X, o.Position.Y, float32(o.Width), float32(o.Height))
+		dest := rl.NewRectangle(float32(o.Position.X), float32(o.Position.Y), float32(o.Width), float32(o.Height))
 
 		if DebugMode && o.DebugVisible {
 			c := o.GetAABB(o)
